+++ exec/Kathu/App/Tools/ToolMode.hs
@@ -3,9 +3,23 @@
 module Kathu.App.Tools.ToolMode where
 
 import Apecs
+import Linear.V2                (V2(..))
+
+import Kathu.App.Graphics.Image (ImageID)
+import Kathu.World.Tile
+
+data TilePlacerState = TilePlacerState
+    { selectedTile :: !(Tile ImageID)
+    , lastPlacePos :: !(Maybe (V2 Int))
+    }
+
+mkTilePlacerState :: Tile ImageID -> TilePlacerState
+mkTilePlacerState t = TilePlacerState t Nothing
 
 -- | A global component that determines whether normal camera rules are overriden and a mouse-based movie should be used
-data ToolMode = NoTool | TilePlacer deriving (Show, Eq)
+data ToolMode
+    = NoTool
+    | TilePlacer TilePlacerState
 
 instance Semigroup ToolMode where (<>) = mappend
 instance Monoid ToolMode where mempty = NoTool