+++ exec/Kathu/Editor/Types.hs
@@ -14,8 +14,8 @@ type MenuBarDescription = [(Text, [(Text, Gtk.MenuItemActivateCallback)])]
 type EditableProperty a = a -> IO ()
 
 data WSEditState = WSEditState
-    { worldspaceRef     :: !(IORef (WorldSpace ImageID))
-    , wsProperties      :: ![EditableProperty (WorldSpace ImageID)]
+    { worldspaceRef :: !(IORef (WorldSpace ImageID))
+    , wsProperties  :: ![EditableProperty (WorldSpace ImageID)]
     }
 
 data EditorState = EditorState